Support Kernel.#Complex, Float#div, Complex#abs, #rect.
